refactor(davinci): share dynamic prop array emit
