+++ crates/vize_s1_to_s2/src/emit.rs
@@ -86,6 +86,7 @@ mod prefix;
 mod props;
 mod props_bind;
 mod props_class;
+mod props_dynamic;
 mod props_object;
 mod props_object_merge;
 mod props_static;
